requirements.txt
django
google-generativeai
openai-whisper
torch
.env